My Code So Far. READ DESCRIPTION
All Chess Pieces can now move the way they are supposed to. Functions
are commented.. Started working on the check() function,  which tells us
if the current move put the opponent in check. Have only finished the
check function for Pawns only, have not finished it for other Chess
Pieces.